+++ src/integrations/text/font.rs
@@ -31,6 +31,7 @@ impl RenderAsset for VelloFont {
 
     fn prepare_asset(
         source_asset: Self::SourceAsset,
+        _asset_id: AssetId<Self::SourceAsset>,
         _param: &mut bevy::ecs::system::SystemParamItem<Self::Param>,
     ) -> Result<Self, bevy::render::render_asset::PrepareAssetError<Self::SourceAsset>> {
         Ok(source_asset)
@@ -55,7 +56,7 @@ impl VelloFont {
 
             LOCAL_LAYOUT_CONTEXT.with_borrow_mut(|layout_context| {
                 let mut builder =
-                    layout_context.ranged_builder(font_context, &text_section.value, 1.0);
+                    layout_context.ranged_builder(font_context, &text_section.value, 1.0, true);
 
                 apply_font_styles(&mut builder, text_section);
                 apply_variable_axes(&mut builder, &text_section.style.font_axes);
@@ -100,7 +101,7 @@ impl VelloFont {
 
             LOCAL_LAYOUT_CONTEXT.with_borrow_mut(|layout_context| {
                 let mut builder =
-                    layout_context.ranged_builder(font_context, &text_section.value, 1.0);
+                    layout_context.ranged_builder(font_context, &text_section.value, 1.0, true);
 
                 apply_font_styles(&mut builder, text_section);
                 apply_variable_axes(&mut builder, &text_section.style.font_axes);