+++ src/lib.rs
@@ -3,7 +3,10 @@
 //! An integration to render SVG and Lottie assets in Bevy with Vello.
 
 use crate::prelude::*;
-use bevy::prelude::*;
+use bevy::{
+    prelude::*,
+    render::view::{self, VisibilityClass},
+};
 
 mod plugin;
 pub use plugin::VelloPlugin;
@@ -52,11 +55,14 @@ pub struct VelloSceneBundle {
     pub transform: Transform,
     /// User indication of whether an entity is visible. Propagates down the entity hierarchy.
     pub visibility: Visibility,
+    /// A bucket into which we group entities for the purposes of visibility.
+    pub visibility_class: VisibilityClass,
 }
 
 /// A simple newtype component wrapper for [`vello::Scene`] for rendering.
 #[derive(Component, Default, Clone, Deref, DerefMut)]
-#[require(Transform, Visibility)]
+#[require(Transform, Visibility, VisibilityClass)]
+#[component(on_add = view::add_visibility_class::<VelloScene>)]
 pub struct VelloScene(Box<vello::Scene>);
 
 impl VelloScene {